# backend/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DukaPay API",
    "version": "1.0.0",
    "description": "Backend API for DukaPay lending, scoring, remittance, and indexer flows."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Legacy API base path"
    },
    {
      "url": "/api/v1",
      "description": "Versioned API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/agents/my-assignments": {
      "get": {
        "summary": "List borrower wallets assigned to the authenticated agent",
        "description": "Returns the borrowers the authenticated agent (or admin) is scoped to under tenant isolation (the `agents:view-assigned` scope). Agents can read but not modify the data of these borrowers on the loan, score and remittance endpoints, enforced both by the RBAC middleware and by RLS.\n",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "agentPublicKey",
                        "assignments"
                      ],
                      "properties": {
                        "agentPublicKey": {
                          "type": "string"
                        },
                        "assignments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "borrowerPublicKey",
                              "createdAt"
                            ],
                            "properties": {
                              "borrowerPublicKey": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token."
          },
          "403": {
            "description": "Requires the agent role."
          }
        }
      }
    },
    "/agents/{agentPublicKey}/assignees": {
      "post": {
        "summary": "Assign a borrower to an agent (admin only)",
        "description": "Grants an agent read access to a borrower's data. Idempotent: assigning the same pair twice returns the existing assignment. Admin only.\n",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentPublicKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent's Stellar public key receiving the assignment"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "borrowerPublicKey"
                ],
                "properties": {
                  "borrowerPublicKey": {
                    "type": "string",
                    "description": "Borrower's Stellar public key to assign"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created (or already present)."
          },
          "400": {
            "description": "Invalid public key."
          },
          "401": {
            "description": "Missing or invalid Bearer token."
          },
          "403": {
            "description": "Requires the admin role."
          }
        }
      }
    },
    "/agents/{agentPublicKey}/assignees/{borrowerPublicKey}": {
      "delete": {
        "summary": "Revoke a borrower assignment from an agent (admin only)",
        "description": "Removes the agent's read access to the borrower. Admin only.\n",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentPublicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "borrowerPublicKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment revoked."
          },
          "400": {
            "description": "Invalid public key."
          },
          "401": {
            "description": "Missing or invalid Bearer token."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Assignment not found."
          }
        }
      }
    },
...
  },
...
}